Add docs and some improvements
