+++ src/lib/constants.ts
@@ -1,6 +1,16 @@
 import packageJson from '../../package.json';
 
 /**
+ * Array of navigation items for the website (i.e. paths/pages to navigate to).
+ */
+export const navItems = [
+    {name: 'Home', path: '/'},
+    {name: 'Work', path: '/work'},
+    {name: 'Projects', path: '/projects'},
+    {name: 'Blog', path: '/blog'}
+]
+
+/**
  * Temporary mapping for tech stack icons.
  */
 export const techStackMap: Record<string, string> = {
